refactor: refactor routes [slug] into [author][slug]
