+++ src/routes/blog/[author]/[slug]/+page.svelte
@@ -1,13 +1,13 @@
 <script lang="ts">
-	import type { PageData } from '../../../../.svelte-kit/types/src/routes/blog/[slug]/$types';
-	import type { ISafeBlogAuthor } from '../../../types/Blog';
+	import type { PageData } from './$types';
+	import type { ISafeBlogAuthor } from '../../../../types/Blog';
 	import { onMount } from 'svelte';
 	import { themeChange } from 'theme-change';
-	import type UserProfile from '../../../types/UserProfile';
-	import DetailBlogBody from '../../../components/body/DetailBlogBody.svelte';
-	import MainHeader from '../../../components/header/MainHeader.svelte';
-	import MetaHead from '../../../components/meta/MetaHead.svelte';
-	import SchemaHead from '../../../components/meta/SchemaHead.svelte';
+	import type UserProfile from '../../../../types/UserProfile';
+	import DetailBlogBody from '../../../../components/body/DetailBlogBody.svelte';
+	import MainHeader from '../../../../components/header/MainHeader.svelte';
+	import MetaHead from '../../../../components/meta/MetaHead.svelte';
+	import SchemaHead from '../../../../components/meta/SchemaHead.svelte';
 
 	export let data: PageData;
 
@@ -28,7 +28,7 @@
 <!-- META TAGs -->
 <MetaHead
 	title={`${blog.Title} by ${blog.Author.Username} | Blog | Resqiar.com`}
-	url={`https://resqiar.com/blog/${blog.ID}`}
+	url={`https://resqiar.com/blog/${blog.Author.Username}/${blog.Slug}`}
 	description={blog.Summary}
 	imageURL={blog.CoverURL}
 />
